feat: implemented processing log operation with rest assurance, and created tests for it Now it is possible log some information during request/response and wrapped request/response will be returned after that: ``` request.log().all().response() request.log().body().request().get(url) request.log().body().get(url) ```
