+++ serenity-rest-assured/src/test/groovy/net/serenitybdd/rest/staging/requests/WhenExecutingDeleteRequestFromExpectation.groovy
@@ -1,4 +1,4 @@
-package net.serenitybdd.rest.staging
+package net.serenitybdd.rest.staging.requests
 
 import com.github.tomakehurst.wiremock.client.WireMock
 import com.github.tomakehurst.wiremock.junit.WireMockRule
@@ -15,12 +15,13 @@ import static com.github.tomakehurst.wiremock.client.WireMock.matching
 import static com.github.tomakehurst.wiremock.client.WireMock.stubFor
 import static com.github.tomakehurst.wiremock.client.WireMock.urlMatching
 
+
 /**
  * User: YamStranger
- * Date: 3/14/16
+ * Date: 3/30/16
  * Time: 9:57 AM
  */
-class WhenExecutingGetRequestFromExpectation extends Specification {
+class WhenExecutingDeleteRequestFromExpectation extends Specification {
 
     @Rule
     def WireMockRule wire = new WireMockRule(0);
@@ -33,15 +34,15 @@ class WhenExecutingGetRequestFromExpectation extends Specification {
         }
     },)
 
-    def "should return wrapped response during GET by URL called from expectation"() {
+    def "should return wrapped response during DELETE by URL called from expectation"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>7</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/get/creature"
+            def path = "/test/delete/creature"
             def url = "$base$path"
-            stubFor(WireMock.get(urlMatching("$path.*"))
+            stubFor(WireMock.delete(urlMatching("$path.*"))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(506)
@@ -52,20 +53,20 @@ class WhenExecutingGetRequestFromExpectation extends Specification {
                 statusCode(506).
                 body(Matchers.equalTo(body))
         and: "executing expectation"
-            def response = expectation.when().get(url);
+            def response = expectation.when().delete(url);
         then: "created response should be decorated"
             response instanceof ResponseDecorated
     }
 
-    def "should return wrapped response during GET by URL called from expectation with parameters"() {
+    def "should return wrapped response during DELETE  by URL called from expectation with parameters"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>7</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/get/creature"
+            def path = "/test/delete/creature"
             def url = "$base$path"
-            stubFor(WireMock.get(urlMatching("$path.*"))
+            stubFor(WireMock.delete(urlMatching("$path.*"))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(506)
@@ -78,7 +79,7 @@ class WhenExecutingGetRequestFromExpectation extends Specification {
                 statusCode(506).
                 body(Matchers.equalTo(body)).
                 when().
-                get(url);
+                delete(url);
         then: "created response should be decorated"
             response instanceof ResponseDecorated
     }
