+++ serenity-rest-assured/src/test/groovy/net/serenitybdd/rest/staging/requests/WhenExecutingGetRequest.groovy
@@ -1,4 +1,4 @@
-package net.serenitybdd.rest.staging
+package net.serenitybdd.rest.staging.requests
 
 import com.github.tomakehurst.wiremock.client.WireMock
 import com.github.tomakehurst.wiremock.junit.WireMockRule
@@ -20,10 +20,10 @@ import static com.github.tomakehurst.wiremock.client.WireMock.urlMatching
 
 /**
  * User: YamStranger
- * Date: 3/14/16
+ * Date: 3/30/16
  * Time: 9:57 AM
  */
-class WhenExecutingHeadRequest extends Specification {
+class WhenExecutingGetRequest extends Specification {
 
     @Rule
     def WireMockRule wire = new WireMockRule(0);
@@ -44,9 +44,9 @@ class WhenExecutingHeadRequest extends Specification {
                 "<value>1</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/head/levels"
+            def path = "/test/get/levels"
             def url = "$base$path"
-            stubFor(WireMock.head(urlMatching(path))
+            stubFor(WireMock.get(urlMatching(path))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(200)
@@ -54,137 +54,137 @@ class WhenExecutingHeadRequest extends Specification {
                 .withBody(body)));
         when: "creating new request and making request"
             def RequestSender sender = given(request, response)
-            def generated = sender.head(url)
+            def generated = sender.get(url)
         then: "created response should be decorated"
             generated instanceof ResponseDecorated
     }
 
-    def "should return wrapped response during HEAD by URL called from request"() {
+    def "should return wrapped response during GET by URL called from request"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>2</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/head/keyboard"
+            def path = "/test/get/keyboard"
             def url = "$base$path"
-            stubFor(WireMock.head(urlMatching(path))
+            stubFor(WireMock.get(urlMatching(path))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(256)
                 .withHeader("Content-Type", "application/xml")
                 .withBody(body)));
-        when: "creating new request and making head request"
-            def response = given().head(url)
+        when: "creating new request and making get request"
+            def response = given().get(url)
         then: "created response should be decorated"
             response instanceof ResponseDecorated
         and: "returned status should be correct"
             response.then().statusCode(256)
     }
 
-    def "should return wrapped response during HEAD by URL with MAP parameters called from request"() {
+    def "should return wrapped response during GET by URL with MAP parameters called from request"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>3</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/head/house"
+            def path = "/test/get/house"
             def url = "$base$path"
-            stubFor(WireMock.head(urlMatching("$path.*"))
+            stubFor(WireMock.get(urlMatching("$path.*"))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(700)
                 .withHeader("Content-Type", "application/xml")
                 .withBody(body)));
-        when: "creating new request and making head request"
-            def response = given().head("$url?status={status}", ["status": "available"])
+        when: "creating new request and making get request"
+            def response = given().get("$url?status={status}", ["status": "available"])
         then: "created response should be decorated"
             response instanceof ResponseDecorated
         and: "returned status should be correct"
             response.then().statusCode(700)
     }
 
-    def "should return wrapped response during HEAD by URL with array parameters called from request"() {
+    def "should return wrapped response during GET by URL with array parameters called from request"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>4</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/head/pet"
+            def path = "/test/get/pet"
             def url = "$base$path"
-            stubFor(WireMock.head(urlMatching("$path.*"))
+            stubFor(WireMock.get(urlMatching("$path.*"))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(845)
                 .withHeader("Content-Type", "application/xml")
                 .withBody(body)));
-        when: "creating new request and making head request"
-            def response = given().head("$url?status={status}", "available")
+        when: "creating new request and making get request"
+            def response = given().get("$url?status={status}", "available")
         then: "created response should be decorated"
             response instanceof ResponseDecorated
         and: "returned status should be correct"
             response.then().statusCode(845)
     }
 
-    def "should return wrapped response during HEAD by URL called from response"() {
+    def "should return wrapped response during GET by URL called from response"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>5</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/head/child"
+            def path = "/test/get/child"
             def url = "$base$path"
-            stubFor(WireMock.head(urlMatching(path))
+            stubFor(WireMock.get(urlMatching(path))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(945)
                 .withHeader("Content-Type", "application/xml")
                 .withBody(body)));
-        when: "creating new request and making head request"
-            def response = expect().head(url)
+        when: "creating new request and making get request"
+            def response = expect().get(url)
         then: "created response should be decorated"
             response instanceof ResponseDecorated
         and: "returned status should be correct"
             response.then().statusCode(945)
     }
 
-    def "should return wrapped response during HEAD by URL with MAP parameters called from response"() {
+    def "should return wrapped response during GET by URL with MAP parameters called from response"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>6</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/head/book"
+            def path = "/test/get/book"
             def url = "$base$path"
-            stubFor(WireMock.head(urlMatching("$path.*"))
+            stubFor(WireMock.get(urlMatching("$path.*"))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(203)
                 .withHeader("Content-Type", "application/xml")
                 .withBody(body)));
-        when: "creating new request and making head request"
-            def response = expect().head("$url?status={status}", ["status": "available"])
+        when: "creating new request and making get request"
+            def response = expect().get("$url?status={status}", ["status": "available"])
         then: "created response should be decorated"
             response instanceof ResponseDecorated
         and: "returned status should be correct"
             response.then().statusCode(203)
     }
 
-    def "should return wrapped response during HEAD by URL with array parameters called from response"() {
+    def "should return wrapped response during GET by URL with array parameters called from response"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>7</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/head/creature"
+            def path = "/test/get/creature"
             def url = "$base$path"
-            stubFor(WireMock.head(urlMatching("$path.*"))
+            stubFor(WireMock.get(urlMatching("$path.*"))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(506)
                 .withHeader("Content-Type", "application/xml")
                 .withBody(body)));
-        when: "creating new request and making head request"
-            def response = expect().head("$url?status={status}", "available")
+        when: "creating new request and making get request"
+            def response = expect().get("$url?status={status}", "available")
         then: "created response should be decorated"
             response instanceof ResponseDecorated
         and: "returned status should be correct"
