+++ serenity-rest-assured/src/test/groovy/net/serenitybdd/rest/staging/requests/WhenExecutingOptionsRequestFromExpectation.groovy
@@ -1,4 +1,4 @@
-package net.serenitybdd.rest.staging
+package net.serenitybdd.rest.staging.requests
 
 import com.github.tomakehurst.wiremock.client.WireMock
 import com.github.tomakehurst.wiremock.junit.WireMockRule
@@ -15,13 +15,12 @@ import static com.github.tomakehurst.wiremock.client.WireMock.matching
 import static com.github.tomakehurst.wiremock.client.WireMock.stubFor
 import static com.github.tomakehurst.wiremock.client.WireMock.urlMatching
 
-
 /**
  * User: YamStranger
- * Date: 3/14/16
+ * Date: 3/30/16
  * Time: 9:57 AM
  */
-class WhenExecutingDeleteRequestFromExpectation extends Specification {
+class WhenExecutingOptionsRequestFromExpectation extends Specification {
 
     @Rule
     def WireMockRule wire = new WireMockRule(0);
@@ -34,15 +33,15 @@ class WhenExecutingDeleteRequestFromExpectation extends Specification {
         }
     },)
 
-    def "should return wrapped response during DELETE by URL called from expectation"() {
+    def "should return wrapped response during OPTIONS by URL called from expectation"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>7</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/delete/creature"
+            def path = "/test/options/creature"
             def url = "$base$path"
-            stubFor(WireMock.delete(urlMatching("$path.*"))
+            stubFor(WireMock.options(urlMatching("$path.*"))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(506)
@@ -53,20 +52,20 @@ class WhenExecutingDeleteRequestFromExpectation extends Specification {
                 statusCode(506).
                 body(Matchers.equalTo(body))
         and: "executing expectation"
-            def response = expectation.when().delete(url);
+            def response = expectation.when().options(url);
         then: "created response should be decorated"
             response instanceof ResponseDecorated
     }
 
-    def "should return wrapped response during DELETE  by URL called from expectation with parameters"() {
+    def "should return wrapped response during OPTIONS by URL called from expectation with parameters"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>7</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/delete/creature"
+            def path = "/test/options/creature"
             def url = "$base$path"
-            stubFor(WireMock.delete(urlMatching("$path.*"))
+            stubFor(WireMock.options(urlMatching("$path.*"))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(506)
@@ -79,7 +78,7 @@ class WhenExecutingDeleteRequestFromExpectation extends Specification {
                 statusCode(506).
                 body(Matchers.equalTo(body)).
                 when().
-                delete(url);
+                options(url);
         then: "created response should be decorated"
             response instanceof ResponseDecorated
     }
