+++ serenity-rest-assured/src/test/groovy/net/serenitybdd/rest/staging/requests/WhenExecutingPutRequest.groovy
@@ -1,4 +1,4 @@
-package net.serenitybdd.rest.staging
+package net.serenitybdd.rest.staging.requests
 
 import com.github.tomakehurst.wiremock.client.WireMock
 import com.github.tomakehurst.wiremock.junit.WireMockRule
@@ -20,10 +20,10 @@ import static com.github.tomakehurst.wiremock.client.WireMock.urlMatching
 
 /**
  * User: YamStranger
- * Date: 3/14/16
+ * Date: 3/30/16
  * Time: 9:57 AM
  */
-class WhenExecutingPatchRequest extends Specification {
+class WhenExecutingPutRequest extends Specification {
 
     @Rule
     def WireMockRule wire = new WireMockRule(0);
@@ -44,9 +44,9 @@ class WhenExecutingPatchRequest extends Specification {
                 "<value>1</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/patch/levels"
+            def path = "/test/put/levels"
             def url = "$base$path"
-            stubFor(WireMock.patch(urlMatching(path))
+            stubFor(WireMock.put(urlMatching(path))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(200)
@@ -54,137 +54,137 @@ class WhenExecutingPatchRequest extends Specification {
                 .withBody(body)));
         when: "creating new request and making request"
             def RequestSender sender = given(request, response)
-            def generated = sender.patch(url)
+            def generated = sender.put(url)
         then: "created response should be decorated"
             generated instanceof ResponseDecorated
     }
 
-    def "should return wrapped response during PATCH by URL called from request"() {
+    def "should return wrapped response during PUT by URL called from request"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>2</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/patch/keyboard"
+            def path = "/test/put/keyboard"
             def url = "$base$path"
-            stubFor(WireMock.patch(urlMatching(path))
+            stubFor(WireMock.put(urlMatching(path))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(256)
                 .withHeader("Content-Type", "application/xml")
                 .withBody(body)));
-        when: "creating new request and making patch request"
-            def response = given().patch(url)
+        when: "creating new request and making put request"
+            def response = given().put(url)
         then: "created response should be decorated"
             response instanceof ResponseDecorated
         and: "returned status should be correct"
             response.then().statusCode(256)
     }
 
-    def "should return wrapped response during PATCH by URL with MAP parameters called from request"() {
+    def "should return wrapped response during PUT by URL with MAP parameters called from request"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>3</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/patch/house"
+            def path = "/test/put/house"
             def url = "$base$path"
-            stubFor(WireMock.patch(urlMatching("$path.*"))
+            stubFor(WireMock.put(urlMatching("$path.*"))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(700)
                 .withHeader("Content-Type", "application/xml")
                 .withBody(body)));
-        when: "creating new request and making patch request"
-            def response = given().patch("$url?status={status}", ["status": "available"])
+        when: "creating new request and making put request"
+            def response = given().put("$url?status={status}", ["status": "available"])
         then: "created response should be decorated"
             response instanceof ResponseDecorated
         and: "returned status should be correct"
             response.then().statusCode(700)
     }
 
-    def "should return wrapped response during PATCH by URL with array parameters called from request"() {
+    def "should return wrapped response during PUT by URL with array parameters called from request"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>4</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/patch/pet"
+            def path = "/test/put/pet"
             def url = "$base$path"
-            stubFor(WireMock.patch(urlMatching("$path.*"))
+            stubFor(WireMock.put(urlMatching("$path.*"))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(845)
                 .withHeader("Content-Type", "application/xml")
                 .withBody(body)));
-        when: "creating new request and making patch request"
-            def response = given().patch("$url?status={status}", "available")
+        when: "creating new request and making put request"
+            def response = given().put("$url?status={status}", "available")
         then: "created response should be decorated"
             response instanceof ResponseDecorated
         and: "returned status should be correct"
             response.then().statusCode(845)
     }
 
-    def "should return wrapped response during PATCH by URL called from response"() {
+    def "should return wrapped response during PUT by URL called from response"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>5</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/patch/child"
+            def path = "/test/put/child"
             def url = "$base$path"
-            stubFor(WireMock.patch(urlMatching(path))
+            stubFor(WireMock.put(urlMatching(path))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(945)
                 .withHeader("Content-Type", "application/xml")
                 .withBody(body)));
-        when: "creating new request and making patch request"
-            def response = expect().patch(url)
+        when: "creating new request and making put request"
+            def response = expect().put(url)
         then: "created response should be decorated"
             response instanceof ResponseDecorated
         and: "returned status should be correct"
             response.then().statusCode(945)
     }
 
-    def "should return wrapped response during PATCH by URL with MAP parameters called from response"() {
+    def "should return wrapped response during PUT by URL with MAP parameters called from response"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>6</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/patch/book"
+            def path = "/test/put/book"
             def url = "$base$path"
-            stubFor(WireMock.patch(urlMatching("$path.*"))
+            stubFor(WireMock.put(urlMatching("$path.*"))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(203)
                 .withHeader("Content-Type", "application/xml")
                 .withBody(body)));
-        when: "creating new request and making patch request"
-            def response = expect().patch("$url?status={status}", ["status": "available"])
+        when: "creating new request and making put request"
+            def response = expect().put("$url?status={status}", ["status": "available"])
         then: "created response should be decorated"
             response instanceof ResponseDecorated
         and: "returned status should be correct"
             response.then().statusCode(203)
     }
 
-    def "should return wrapped response during PATCH by URL with array parameters called from response"() {
+    def "should return wrapped response during PUT by URL with array parameters called from response"() {
         given: "configured access point"
             def body = "<root>" +
                 "<value>7</value>" +
                 "</root>"
             def base = "http://localhost:${wire.port()}"
-            def path = "/test/patch/creature"
+            def path = "/test/put/creature"
             def url = "$base$path"
-            stubFor(WireMock.patch(urlMatching("$path.*"))
+            stubFor(WireMock.put(urlMatching("$path.*"))
                 .withRequestBody(matching(".*"))
                 .willReturn(aResponse()
                 .withStatus(506)
                 .withHeader("Content-Type", "application/xml")
                 .withBody(body)));
-        when: "creating new request and making patch request"
-            def response = expect().patch("$url?status={status}", "available")
+        when: "creating new request and making put request"
+            def response = expect().put("$url?status={status}", "available")
         then: "created response should be decorated"
             response instanceof ResponseDecorated
         and: "returned status should be correct"
